Breaking up the handler module
