+++ src/app/queue/handlers/onNewFeedback/layoutNotifyMail.ts
@@ -16,90 +16,25 @@
 // this program. If not, see <http://www.gnu.org/licenses/>.
 //
 
-import {
-	DoneCallback,
-	Job,
-} from 'kue';
-
 import Article from 'base/Article';
 import ArticleItem from 'base/ArticleItem';
 import Feedback from 'base/Feedback';
 import FeedbackItem from 'base/FeedbackItem';
 import MailTemplate from 'app/template/MailTemplate';
-import Website from 'base/Website';
-
-import { ObjectID } from 'app/db';
-
-import {
-	feedbackService,
-	localizationService,
-	templateService,
-	websiteService,
-} from 'app/services';
 
 import {
 	format,
 	ItemFormatPayload,
 } from 'app/mail/layout/FeedbackNotifyLayout';
 
-import SendGridMailer from 'app/mail/sendgrid/SendGridMailer';
-
-import * as app from 'app/util/applib';
-
-const __ = localizationService.translate;
-const log = app.createLog();
-
-export default function(job : Job, done : DoneCallback) : Promise <void> {
-	if (!job.data.ID) {
-		log('Feedback "ID" not found in job data');
-		return Promise.resolve();
-	}
-
-	log(`Received new feedback event for ID ${job.data.ID}`);
-
-	let feedback : Feedback;
-	let template : MailTemplate;
-	let website : Website;
-
-	return new Promise <void> ((resolve, reject) => {
-		Promise.all([
-			feedbackService.getByID(job.data.ID),
-			templateService.getFeedbackNotifyTemplate(),
-		])
-		.spread((f : Feedback, t : MailTemplate) => {
-			feedback = f;
-			template = t;
-
-			return websiteService.getByID(feedback.article.website);
-		})
-		.then((w : Website) => {
-			website = w;
-			log(app.inspect(website));
-
-			// return layoutNotifyMail ...
-		})
-
-		// 	layoutNotifyMail)
-		// .then((htmlMailContent : string) => {
-		// 	const recipients = getRecipients
-		// 	log(htmlMailContent);
-		// 	return SendGridMailer('philipp@sol.no', 'Leserkritikk', htmlMailContent);
-		// 	// Here be the sending of the mail. This is solved in feature/mail-sendgrid
-		// })
-		.then(() => {
-			done();
-			resolve();
-		})
-		.catch(reject);
-	});
-}
+import { translate as __ } from 'app/services/localization';
 
 const cssFeedbackItemBox = [
 	'padding: 0.8em',
 	'margin-bottom: 0.5em',
 ].join(';');
 
-function layoutNotifyMail(feedback : Feedback, template : MailTemplate) : Promise <any> {
+export default function (feedback : Feedback, template : MailTemplate) : Promise <any> {
 	let dtxt = '';
 
 	feedback.items.forEach((fItem : FeedbackItem, fIndex : number) => {
@@ -154,23 +89,3 @@ function debugInfo(feedback : Feedback) : string {
 		`article version = ${feedback.article.version}`,
 	].join('<br/>');
 }
-
-// Only for development
-
-/*
-import { writeFileSync } from 'fs';
-import { spawn } from 'child_process';
-
-function notifyBrowser(html : string) {
-	const testPath = '/tmp/mailtest.html';
-
-	writeFileSync(testPath, html, {
-		flag: 'w',
-		mode: 0o644,
-	});
-
-	spawn('/usr/bin/qupzilla', [ '-c', `file://${testPath}` ], {
-		detached: true,
-	});
-}
-*/
